# servicios/transparencia/src/main/resources/openapi/transparencia.yaml
...
openapi: 3.1.0
info:
  title: transparencia
  version: 1.0.0
  description: >-
    Servicio transparencia. Un caso de uso = una operacion, con el codigo CU en el
    operationId. Los importes viajan como CADENA decimal, nunca como number.

    Este servicio existe para que nadie tenga que creernos. Por eso tiene las dos
    unicas familias de rutas sin sesion del sistema: /publico y /verificar.
servers:
  - url: /api/v1
paths:
  /publico/sorteos/{sorteoId}/verificacion:
    get:
      operationId: verificarSorteo              # CU-61
      summary: Verificar publicamente el sorteo
      description: >-
        Ruta PUBLICA, sin sesion. Recomputa el compromiso y el barajado con el mismo
        atomo que uso CU-60 para sortear: una sola implementacion para generar y para
        verificar. Antes del revelado no hay nada que verificar, y ese es el punto.
      parameters:
        - name: sorteoId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Veredicto, con el paquete completo para rehacerlo por fuera
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaVerificacionSorteo' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
...
  /publico/grupos/{grupoId}/verificacion:
    get:
      operationId: verificarCadena              # CU-73
      summary: Verificar la cadena de transparencia
      description: >-
        Ruta PUBLICA, sin sesion. Recorre los bloques y devuelve el primero que falla y
        que componente no cuadra. Un grupo sin bloques no es un error de integridad.
      parameters:
        - name: grupoId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Veredicto de la cadena
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaVerificacionCadena' }
        '422': { $ref: '#/components/responses/ReglaDeNegocio' }
  /publico/grupos/{grupoId}/bloques:
    get:
      operationId: listarBloques                # CU-72
      summary: Los bloques publicados de un grupo
      description: >-
        Ruta PUBLICA, sin sesion. Es el paquete que cualquiera descarga para verificar
        por fuera.
      parameters:
        - name: grupoId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: La cadena, en orden
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BloquePublicado' }
...
components:
...
  responses:
    ReglaDeNegocio:
      description: Una regla del dominio lo impide. El codigo AP-CUnn-nn dice cual.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problema' }
  schemas:
...
    Problema:
      type: object
      additionalProperties: false
      required: [tipo, titulo, estado, codigo]
      properties:
        tipo:    { type: string }
        titulo:  { type: string }
        estado:  { type: integer }
        codigo:  { type: string, pattern: '^AP-CU[0-9]{2,3}-[0-9]{2}$' }
        detalle: { type: [string, 'null'] }
    Hash:
      type: string
      minLength: 64
      maxLength: 64
      description: SHA-256 en hexadecimal minuscula.
...
    SalidaVerificacionSorteo:
      type: object
      additionalProperties: false
      required: [verifica, hashEsperado, hashRecomputado, ordenCoincide, paquete]
      properties:
        verifica:        { type: boolean }
        hashEsperado:    { $ref: '#/components/schemas/Hash' }
        hashRecomputado: { $ref: '#/components/schemas/Hash' }
        ordenCoincide:   { type: boolean }
        primerCupoDiscrepante: { type: [integer, 'null'] }
        paquete:
          type: object
          additionalProperties: false
          required: [semilla, entropias, metodo, cupos]
          properties:
            semilla:   { $ref: '#/components/schemas/Hash' }
            entropias: { type: array, items: { type: string } }
            metodo:    { type: string }
            cupos:     { type: array, items: { type: integer } }

    # ------------------------------------------------------------------ CU-70
...
    BloquePublicado:
      type: object
      additionalProperties: false
      required: [numeroBloque, hashAnterior, raizMerkle, hashBloque, desde, hasta, cantidadEventos]
      properties:
        numeroBloque:    { type: integer, format: int64 }
        hashAnterior:    { $ref: '#/components/schemas/Hash' }
        raizMerkle:      { $ref: '#/components/schemas/Hash' }
        hashBloque:      { $ref: '#/components/schemas/Hash' }
        desde:           { type: string, format: date-time }
        hasta:           { type: string, format: date-time }
        cantidadEventos: { type: integer, minimum: 0 }

    # ------------------------------------------------------------------ CU-73
    SalidaVerificacionCadena:
      type: object
      additionalProperties: false
      required: [integra, bloquesVerificados]
      properties:
        integra:            { type: boolean }
        bloquesVerificados: { type: integer, minimum: 0 }
        primerBloqueFallido: { type: [integer, 'null'], format: int64 }
        componenteFallido:
          type: [string, 'null']
          enum: [HASH_CONTENIDO, HASH_ANTERIOR, HASH_BLOQUE, SECUENCIA, null]
        ultimoSellado: { type: [string, 'null'], format: date-time }

    # ------------------------------------------------------------------ CU-74
...